chore(sdk): Move `RoomListEntry` into the `sliding_sync::view` module.
